[project @ 2005-02-17 13:51:25 by duncan_coutts] Some more doc improvements
Add the ability to exclude certain modules when scanning existing modules.
This is useful so that we do not confuse some Gtk module with a Gdk or Pango
module of the same name.

Allow functions specified in a file to not have code generated for them. It
uses the same format as the apicoverge tool's api.ignore files. Use this to
exclude the *_get_type, *_valist and *_ref / *_unref funtions.