[project @ 2005-04-19 02:07:56 by duncan_coutts] Oops, missed this out of the last commit.
New home of GenericValue deinition until it is eliminated entirely. Comment out
the constructors that are unsafe or in the case of the CChar ones how they are
to be used. Instead of making GenericValue an instance of Storable, define
value(G|S)etGenericValue functions similar to the ones in GValueTypes. Also
convert license to LGPL.