Add support of new floating reference in GObject.
The floating reference needs to be ignored just as in GtkObject.
There are two issues with this patch:
a) it's untested, it might not even compile
b) it should use GTK_CHECK_VERSION when it should use GLIB_CHECK_VERSION
For b) we don't have the appropriate macros yet, and this change will
probably stay the only thing that needs this macro.