Improve the new GObject reference-count update.
I've added new macros to test for the appropriate GLIB verions. This seems
to be the first time that we compile things differently due to different
GLIB versions. I also removed an incorrect comment from GObject:
The semantic of makeGObject and constructGObject stays the same as before.

There's a glitch when building Structs.hsc: cpp complaints that the
GLIB version definitions are duplicated. This is true, but it elludes
me why this doesn't happen for the GTK and PANGO macros, too.