Gtk+2.0 Tutorial Port (part 1)
Start of a port to Gtk2Hs of the Gtk+2.0 tutorial
included in the Gtk documentation.

Note: Chapter 2 is supposed to be an introductory overview,
      as in the original, but for Gtk2Hs. I'm hoping someone
      who knows more will jump in here.
        
The implemented chapters are:
      1.   Contents and copyright, as required by the original authors
      2. To be written by an expert...
      3.   Getting Started
      4.   Packing Widgets
      5.a. Packing Demonstration Program
      5.b. Packing Using Tables
      6.   The Button Widget
      7.   Adjustments, Scale and Range

The examples have been tested with Gtk2Hs-0.9.12 
(the first ones with .11) on Fedora 6. 

The html files have been taken from the original Gtk+2.0 tutorial,
processed in Open Office Writer html mode, and then cleaned up
with Screem and tidy. There are still some unreferred name tags left
from the original.